Fix return type for getSubscriptionById on the WebhookController
I think the return type here was too strict, since the `find` method could return null, e.g. if a subscription has been deleted, and that would cause an error.

Also, on the `cancelSubscription` method we are checking if the $subscription is truthy, which means that is a valid case.